linear algorithm

+++ 0x1E-search_algorithms/0-linear.c
@@ -13,7 +13,7 @@
 
 int linear_search(int *array, size_t size, int value)
 {
-	unsigned int i;
+	long unsigned int i;
 
 	if (!array)
 		return (-1);
